Rename: modal 폴더로 파일 이동, 연관 파일 경로 수정
